refactor: migrate Pages/App to Pages/Applications & add parseAllSearchParamsForUseUrlFilters
