+++ src/Common/Hooks/useUrlFilters/index.ts
@@ -16,3 +16,4 @@
 
 export { default as useUrlFilters } from './useUrlFilters'
 export type { UseUrlFiltersProps, UseUrlFiltersReturnType } from './types'
+export { parseAllSearchParamsForUseUrlFilters } from './utils'
