+++ src/Common/Hooks/useUrlFilters/utils.tsx
@@ -3,3 +3,9 @@ export const setItemInLocalStorageIfKeyExists = (localStorageKey: string, value:
         localStorage.setItem(localStorageKey, value)
     }
 }
+
+export const parseAllSearchParamsForUseUrlFilters = <T,>(params: URLSearchParams) =>
+    Array.from(params.entries()).reduce((acc, [key, value]) => {
+        acc[key] = value
+        return acc
+    }, {} as T)
