--- src/Pages/App/AppConfiguration/DeploymentTemplate/index.ts
@@ -1,1 +1,0 @@
-export * from './GUIView'
