--- src/Pages/App/AppConfiguration/index.ts
@@ -1,1 +1,0 @@
-export * from './DeploymentTemplate'
