--- src/Pages/App/Trigger/index.ts
@@ -1,3 +1,0 @@
-export { triggerCDNode } from './service'
-export { getAPIOptionsWithTriggerTimeout } from './utils'
-export type { TriggerCDNodeServiceProps, TriggerCDPipelinePayloadType } from './types'
