--- src/Pages/App/index.ts
@@ -1,3 +1,0 @@
-export * from './AppConfiguration'
-export * from './Trigger'
-export * from './AppDetails'
