+++ src/Pages/Applications/DevtronApps/Details/AppConfigurations/DeploymentTemplate/GUIView/utils.ts
