+++ src/Pages/Applications/DevtronApps/Details/AppConfigurations/DeploymentTemplate/index.ts
@@ -2,3 +2,4 @@ export * from './types'
 export * from './service'
 export { default as DTApplicationMetricsFormField } from './DTApplicationMetricsFormField'
 export { GET_RESOLVED_DEPLOYMENT_TEMPLATE_EMPTY_RESPONSE } from './constants'
+export * from './GUIView'
