+++ src/Pages/Applications/DevtronApps/Details/CDPipeline/constants.ts
