+++ src/Pages/Applications/DevtronApps/Details/CDPipeline/services.ts
@@ -1,9 +1,13 @@
 import { MutableRefObject } from 'react'
-
 import { getIsRequestAborted, post } from '@Common/Api'
 import { ROUTES } from '@Common/Constants'
 import { getUrlWithSearchParams, showError } from '@Common/Helper'
 import { UploadFileDTO, UploadFileProps } from '@Shared/types'
+import { DeploymentNodeType } from '@Common/Types'
+import { DeploymentWithConfigType } from '@Shared/constants'
+import { TriggerCDNodeServiceProps, TriggerCDPipelinePayloadType } from './types'
+import { getAPIOptionsWithTriggerTimeout } from './utils'
+import { STAGE_MAP } from './constants'
 
 export const uploadCDPipelineFile = async ({
     file,
@@ -38,4 +42,39 @@ export const uploadCDPipelineFile = async ({
         }
         throw err
     }
+}
+
+export const triggerCDNode = ({
+    pipelineId,
+    ciArtifactId,
+    appId,
+    stageType,
+    deploymentWithConfig,
+    wfrId,
+    runtimeParamsPayload,
+    abortControllerRef,
+}: TriggerCDNodeServiceProps) => {
+    const areRuntimeParamsConfigured =
+        runtimeParamsPayload && (stageType === DeploymentNodeType.POSTCD || stageType === DeploymentNodeType.PRECD)
+
+    const request: TriggerCDPipelinePayloadType = {
+        pipelineId,
+        appId,
+        ciArtifactId,
+        cdWorkflowType: STAGE_MAP[stageType],
+        ...(areRuntimeParamsConfigured && runtimeParamsPayload),
+    }
+
+    if (deploymentWithConfig) {
+        request.deploymentWithConfig =
+            deploymentWithConfig === DeploymentWithConfigType.LAST_SAVED_CONFIG
+                ? deploymentWithConfig
+                : DeploymentWithConfigType.SPECIFIC_TRIGGER_CONFIG
+
+        if (deploymentWithConfig !== DeploymentWithConfigType.LAST_SAVED_CONFIG) {
+            request.wfrIdForDeploymentWithSpecificTrigger = wfrId
+        }
+    }
+
+    return post(ROUTES.CD_TRIGGER_POST, request, { ...getAPIOptionsWithTriggerTimeout(), abortControllerRef })
 }
